+++ src/directives.js
@@ -425,10 +425,8 @@ var REMOVE_ATTRIBUTES = {
 angularDirective("ng:bind-attr", function(expression){
   return function(element){
     var lastValue = {};
-    var updateFn = element.data($$update) || noop;
     this.$observe(function(scope){
-      var values = scope.$eval(expression),
-          dirty = noop;
+      var values = scope.$eval(expression);
       for(var key in values) {
         var value = compileBindTemplate(values[key])(scope, element),
             specialName = REMOVE_ATTRIBUTES[lowercase(key)];
@@ -446,10 +444,8 @@ angularDirective("ng:bind-attr", function(expression){
           } else {
             element.attr(key, value);
           }
-          dirty = updateFn;
         }
       }
-      dirty();
     });
   };
 });